Update DCC object attribute names
They decided to capitalize the type names, I needed to update to match
them.

Also there were some new attributes that look useful, so I added them to
the list of things I'm capturing.

Lastly in preparation for another change there was a small refactor
for doing the network request.